requirements.txt
aiogram==3.14.0
python-dotenv==1.0.1
pytz==2024.1
